feat: publish works
