+++ Makefile
@@ -154,13 +154,14 @@ docker:
 	$(DOCKER_BUILDER) build $(DOCKER_BUILD_ARGS) -t $(DOCKER_REGISTRY)/$(DOCKER_REPO)/server:$(VERSION) -f Dockerfile --build-arg LDFLAGS="$(LDFLAGS)" .
 	@echo "✓ Docker image built successfully"
 
-docker-pull-as-latest:
+docker-tag-as-latest:
 	@echo "Pulling and tagging as latest..."
 	docker pull $(DOCKER_REGISTRY)/$(DOCKER_REPO)/server:$(VERSION)
 	docker tag $(DOCKER_REGISTRY)/$(DOCKER_REPO)/server:$(VERSION) $(DOCKER_REGISTRY)/$(DOCKER_REPO)/server:latest
+	docker push $(DOCKER_REGISTRY)/$(DOCKER_REPO)/server:latest
 	@echo "✓ Docker image pulled successfully"
 
-docker-compose-up: docker docker-pull-as-latest
+docker-compose-up: docker docker-tag-as-latest
 	@echo "Starting services with Docker Compose..."
 	docker compose -p agentregistry -f internal/daemon/docker-compose.yml up -d --wait
