+++ internal/cli/remove.go
@@ -6,16 +6,18 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var removeCmd = &cobra.Command{
-	Use:   "remove <resource-type> <resource-name>",
-	Short: "Remove a deployed resource",
-	Long:  `Remove deployed resources (mcp server, skill, agent) from the runtime.`,
-	Args:  cobra.ExactArgs(2),
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Not implemented yet")
-	},
+func newRemoveCommand() *cobra.Command {
+	return &cobra.Command{
+		Use:   "remove <resource-type> <resource-name>",
+		Short: "Remove a deployed resource",
+		Long:  `Remove deployed resources (mcp server, skill, agent) from the runtime.`,
+		Args:  cobra.ExactArgs(2),
+		Run: func(cmd *cobra.Command, args []string) {
+			fmt.Println("Not implemented yet")
+		},
+	}
 }
 
 func init() {
-	rootCmd.AddCommand(removeCmd)
+	rootCmd.AddCommand(newRemoveCommand())
 }
